refactor(backend/copilot): move _extract_cache_creation_tokens to token_tracking
Both service.py and baseline/service.py already import from token_tracking;
moving the helper there eliminates the inline duplication and the comment
explaining it. Test import updated to import from token_tracking directly.
